shifted the integration tests out of main tests into a temp folder
